+++ sdk/storage/storage-file-share/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "js",
   "TagPrefix": "js/storage/storage-file-share",
-  "Tag": "js/storage/storage-file-share_38d5545936"
+  "Tag": "js/storage/storage-file-share_183c6c916d"
 }
