+++ sdk/storage/storage-file-share/src/generatedModels.ts
@@ -41,7 +41,7 @@ import type {
 } from "./generated/src/models";
 
 import type { ShareSetPropertiesHeaders } from "./generated/src/models";
-import { NfsFileMode } from "./models";
+import { FileDownloadResponse, FileSetHttpHeadersHeaders, NfsFileMode } from "./models";
 import type { WithResponse } from "./utils/utils.common";
 
 /** Defines headers for Directory_create operation. */
@@ -533,24 +533,6 @@ export interface FileDownloadHeaders {
 }
 
 /** Contains response data for the download operation. */
-export type FileDownloadResponse = FileDownloadHeaders & {
-  /**
-   * BROWSER ONLY
-   *
-   * The response body as a browser Blob.
-   * Always `undefined` in node.js.
-   */
-  blobBody?: Promise<Blob>;
-  /**
-   * NODEJS ONLY
-   *
-   * The response body as a node.js Readable stream.
-   * Always `undefined` in the browser.
-   */
-  readableStreamBody?: NodeJS.ReadableStream;
-};
-
-/** Contains response data for the download operation. */
 export type FileDownloadResponseModel = WithResponse<FileDownloadResponse, FileDownloadHeaders>;
 
 /** Contains response data for the uploadRangeFromURL operation. */
@@ -578,46 +560,6 @@ export type FileGetRangeListDiffResponse = WithResponse<
   FileGetRangeListHeaders,
   ShareFileRangeList
 >;
-
-/** Defines headers for File_setHttpHeaders operation. */
-export interface FileSetHttpHeadersHeaders {
-  /** The ETag contains a value which represents the version of the file, in quotes. */
-  etag?: string;
-  /** Returns the date and time the directory was last modified. Any operation that modifies the directory or its properties updates the last modified time. Operations on files do not affect the last modified time of the directory. */
-  lastModified?: Date;
-  /** This header uniquely identifies the request that was made and can be used for troubleshooting the request. */
-  requestId?: string;
-  /** Indicates the version of the File service used to execute the request. */
-  version?: string;
-  /** A UTC date/time value generated by the service that indicates the time at which the response was initiated. */
-  date?: Date;
-  /** The value of this header is set to true if the contents of the request are successfully encrypted using the specified algorithm, and false otherwise. */
-  isServerEncrypted?: boolean;
-  /** Key of the permission set for the file. */
-  filePermissionKey?: string;
-  /** Attributes set for the file. */
-  fileAttributes?: string;
-  /** Creation time for the file. */
-  fileCreatedOn?: Date;
-  /** Last write time for the file. */
-  fileLastWriteOn?: Date;
-  /** Change time for the file. */
-  fileChangeOn?: Date;
-  /** The fileId of the directory. */
-  fileId?: string;
-  /** The parent fileId of the directory. */
-  fileParentId?: string;
-  /** NFS only. The mode of the file or directory. */
-  fileMode?: NfsFileMode;
-  /** NFS only. The owner of the file or directory. */
-  owner?: string;
-  /** NFS only. The owning group of the file or directory. */
-  group?: string;
-  /** NFS only. The link count of the file or directory. */
-  linkCount?: number;
-  /** Error Code */
-  errorCode?: string;
-}
 
 /** Contains response data for the setHttpHeaders operation. */
 export type FileSetHTTPHeadersResponse = WithResponse<
@@ -704,7 +646,7 @@ export {
 export {
   FileDownloadResponse as RawFileDownloadResponse,
   FileSetHttpHeadersHeaders as FileSetHTTPHeadersHeaders,
-};
+} from "./models"
 /** Known values of {@link ShareTokenIntent} that the service accepts. */
 export enum KnownShareTokenIntent {
   Backup = "backup",
