+++ sdk/storage/storage-file-share/src/models.ts
@@ -5,6 +5,7 @@ import { truncatedISO8061Date } from "./utils/utils.common";
 import { logger } from "./log";
 import type { FilePermissionFormat, NfsFileType, ShareTokenIntent } from "./generatedModels";
 import type { StoragePipelineOptions } from "./Pipeline";
+import { FileDownloadHeaders } from "./generatedModels";
 
 export interface Metadata {
   [propertyName: string]: string;
@@ -80,6 +81,64 @@ export interface FilePosixProperties {
   /** Optional, NFS only. Type of the file or directory. */
   fileType?: NfsFileType;
 }
+
+/** Defines headers for File_setHttpHeaders operation. */
+export interface FileSetHttpHeadersHeaders {
+  /** The ETag contains a value which represents the version of the file, in quotes. */
+  etag?: string;
+  /** Returns the date and time the directory was last modified. Any operation that modifies the directory or its properties updates the last modified time. Operations on files do not affect the last modified time of the directory. */
+  lastModified?: Date;
+  /** This header uniquely identifies the request that was made and can be used for troubleshooting the request. */
+  requestId?: string;
+  /** Indicates the version of the File service used to execute the request. */
+  version?: string;
+  /** A UTC date/time value generated by the service that indicates the time at which the response was initiated. */
+  date?: Date;
+  /** The value of this header is set to true if the contents of the request are successfully encrypted using the specified algorithm, and false otherwise. */
+  isServerEncrypted?: boolean;
+  /** Key of the permission set for the file. */
+  filePermissionKey?: string;
+  /** Attributes set for the file. */
+  fileAttributes?: string;
+  /** Creation time for the file. */
+  fileCreatedOn?: Date;
+  /** Last write time for the file. */
+  fileLastWriteOn?: Date;
+  /** Change time for the file. */
+  fileChangeOn?: Date;
+  /** The fileId of the directory. */
+  fileId?: string;
+  /** The parent fileId of the directory. */
+  fileParentId?: string;
+  /** NFS only. The mode of the file or directory. */
+  fileMode?: NfsFileMode;
+  /** NFS only. The owner of the file or directory. */
+  owner?: string;
+  /** NFS only. The owning group of the file or directory. */
+  group?: string;
+  /** NFS only. The link count of the file or directory. */
+  linkCount?: number;
+  /** Error Code */
+  errorCode?: string;
+}
+
+/** Contains response data for the download operation. */
+export type FileDownloadResponse = FileDownloadHeaders & {
+  /**
+   * BROWSER ONLY
+   *
+   * The response body as a browser Blob.
+   * Always `undefined` in node.js.
+   */
+  blobBody?: Promise<Blob>;
+  /**
+   * NODEJS ONLY
+   *
+   * The response body as a node.js Readable stream.
+   * Always `undefined` in the browser.
+   */
+  readableStreamBody?: NodeJS.ReadableStream;
+};
 
 export interface FileHttpHeaders {
   /**
